+++ coffeeapp/src/app/app.module.ts
@@ -5,7 +5,7 @@ import {CommonModule} from '@angular/common';
 import {HttpClientModule} from '@angular/common/http';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {ApiService} from './api.service';
+import {ApiService} from './_service';
 import {ConfigComponent} from './config/config.component';
 import {DashboardComponent} from './dashboard/dashboard.component';
 import {ErrorComponent} from './error/error.component';
