+++ coffeeapp/src/app/dashboard/dashboard.component.ts
@@ -1,7 +1,7 @@
 import {Component, ElementRef, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {timer} from 'rxjs';
 import * as moment from 'moment';
-import {ApiService} from '../api.service';
+import {ApiService} from '../_service';
 
 @Component({
   selector: 'app-dashboard',
